requirements/base.txt
Django>=4.2
psycopg2-binary
django-filter
djangorestframework
drf-spectacular 
drf-spectacular-sidecar
django-filter
PyJWT
django-cors-headers
Pillow
channels
daphne
